fix reality check and password
